Fix: bugs, router params, date editing, entry for cents, change item price to Long, refactoring code

+++ app/src/main/java/com/omna/summa/data/local/entity/ShoppingItemEntity.kt
@@ -23,5 +23,5 @@ data class ShoppingItemEntity(
     var name: String,
     var quantity: Double,
     var unit: String,
-    var unitPrice: Double?
+    var unitPrice: Long?
 )
